20230402提交

+++ src/main/java/com/gd/upmms/common/BaseContext.java
@@ -9,13 +9,13 @@ package com.gd.upmms.common;
  */
 public class BaseContext {
 
-    private static ThreadLocal<Long> threadLocal=new InheritableThreadLocal<>();
+    private static ThreadLocal<Integer> threadLocal=new InheritableThreadLocal<>();
 
-    public static void setCurrentId(Long id){
+    public static void setCurrentId(Integer id){
         threadLocal.set(id);
     }
 
-    public static Long getCurrentId(){
+    public static Integer getCurrentId(){
         return threadLocal.get();
     }
 }
